+++ include/main_menu_screen.h
@@ -11,7 +11,7 @@ public:
     TA_ScreenState update() override;
 
 private:
-    const double transitionTime = 5;
+    const float transitionTime = 5;
 
     void updateTitle();
 
@@ -19,7 +19,7 @@ private:
     std::array<std::unique_ptr<TA_MainMenuSection>, TA_MAIN_MENU_MAX> sections;
     TA_MainMenuState state, neededState;
     TA_Sprite titleSprite;
-    double timer = 0;
+    float timer = 0;
 };
 
 #endif // TA_MAIN_MENU_SCREEN_H
