+++ include/objects/bat_robot.h
@@ -7,15 +7,15 @@ class TA_BatRobot : public TA_Object {
 private:
     enum State { STATE_IDLE, STATE_ACTIVE, STATE_ATTACK };
 
-    const double speed = 0.7;
-    const double gravity = -0.1 * speed * speed;
-    const double cooldownTime = 48;
+    const float speed = 0.7;
+    const float gravity = -0.1 * speed * speed;
+    const float cooldownTime = 48;
 
-    double findOptimalSpeed(double deltaY);
+    float findOptimalSpeed(float deltaY);
 
     TA_Point velocity;
     State state = STATE_IDLE;
-    double timer = 0;
+    float timer = 0;
 
 public:
     using TA_Object::TA_Object;
