+++ include/objects/bomb_thrower.h
@@ -5,9 +5,9 @@
 
 class TA_BombThrower : public TA_Object {
 private:
-    const double idleTime = 30;
-    const double walkBackTime = 100;
-    const double speed = 0.7;
+    const float idleTime = 30;
+    const float walkBackTime = 100;
+    const float speed = 0.7;
 
     void updateIdle();
     void updateWalkForward();
@@ -21,20 +21,20 @@ private:
     enum State { STATE_IDLE, STATE_WALK_FORWARD, STATE_WALK_BACK, STATE_ATTACK };
 
     State state = STATE_IDLE;
-    double leftX, rightX;
+    float leftX, rightX;
     bool direction = false;
-    double timer = 0;
+    float timer = 0;
 
 public:
     using TA_Object::TA_Object;
-    void load(TA_Point position, double leftX, double rightX);
+    void load(TA_Point position, float leftX, float rightX);
     bool update() override;
     int getCollisionType() override { return TA_COLLISION_DAMAGE | TA_COLLISION_TARGET; }
 };
 
 class TA_EnemyBomb : public TA_Object {
 private:
-    const double grv = 0.125;
+    const float grv = 0.125;
     const TA_Point topLeft{1, 6}, bottomRight{7, 12};
 
     TA_Point velocity{-1.2, -2.4};
